Tidy up/mark dubious spaced curly quotes
If files use curly quotes, then (unlike with straight quotes) if the quotes have spaces
before and after, it is generally possible to know which is the erroneous one.
Add an option to do this check and to correct or mark them - see below.

Only ambiguous case is single close quote. One of these with spaces before and
after could be a close quote with erroneous space before, but could also be an
apostrophe (e.g. dialect 'orrible) with erroneous space after. Second option switches
between fixing these (if CP knows there is no dialect) or just marking them as
suspect (the default)

Also updated changelog & manual.