Make fcannos.bin platform independent
The file fcannos.bin is a binary Storage of a perl hash, presumably used for speed at
some time in the past when this was an issue for guiprep. The problem was when a
user was on a different platform, e.g. big-endian/little-endian to the developer who
created fcannos.bin. The option to fix Old Englifh then caused an error and
processing stopped.

This has happened several times over the last few years, so it seems sensible to try
to make fcannos.bin platform independent, which I believe it can be if written with
`nstore` instead of `store`.

Two changes therefore: make fwordgen.pl use `nstore`, and include a copy of the
platform-independent version of fcannos.bin

Since on a modern computer fwordgen.pl runs very quickly, a longer term solution
might be to just include the source dictionary file `2of4brif.txt`, and get guiprep
to create its own hash directly from that on startup or when it needs it.

This should be sufficient for now, though.