Force scroll bars on directory lists
When the selected drive contains more directories than will fit in the list,
the scrollbar is sometimes not enabled: it is when the program starts, but
not when the user changes to that drive (reported on Windows).
Forcing the vertical scroll bar to be visible is harmless when it isn't needed.